requirements.txt
asgiref==3.4.1
autopep8==1.6.0
boto3==1.23.10
botocore==1.26.10
Django==3.2.14
django-restframework==0.0.1
django-storages==1.12.3
djangorestframework==3.13.1
jmespath==0.10.0
Pillow==8.4.0
pkg_resources==0.0.0
pycodestyle==2.8.0
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
s3transfer==0.5.2
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
typing_extensions==4.1.1
urllib3==1.26.10
